Add the ability for the bot to react using the reactx command
Pharmacists who want to ignore specific rdzips from a message should react with the number/s corresponding to the order of the attachment/s, then click the "Ignore rdzip attachments" command. The bot will then add its own number reaction to the message, and it will ignore the attachments when indexing. If no number reaction is seen, then the indexer will ignore all attachments in the message.